Added Springboot logic
Added all the Springboot logic to the controllers and the services. Fixed security and added role management in the methods.

+++ src/main/java/it/unicam/ids/Vseet/Model/Services/ItineraryService.java
@@ -1,37 +1,82 @@
 package it.unicam.ids.Vseet.Model.Services;
 
-import it.unicam.ids.Vseet.Model.Entities.Content;
-import it.unicam.ids.Vseet.Model.Entities.ContentCategory;
-import it.unicam.ids.Vseet.Model.Entities.Itinerary;
-import it.unicam.ids.Vseet.Model.Entities.User;
+import it.unicam.ids.Vseet.Model.Entities.*;
 import it.unicam.ids.Vseet.Model.Exceptions.ContentNotFoundException;
+import it.unicam.ids.Vseet.Model.Repositories.ContentRepository;
 import it.unicam.ids.Vseet.Model.Repositories.ItineraryRepository;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
-public class ItineraryService implements SpringService{
-    ItineraryRepository itineraryRepository;
+public class ItineraryService {
+    private final ItineraryRepository itineraryRepository;
+    private final UserService userService;
+    private final ContentRepository contentRepository;
 
-    public ItineraryService(ItineraryRepository itineraryRepository) {
+    public ItineraryService(ItineraryRepository itineraryRepository, ContentRepository contentRepository,
+                            UserService userService) {
         this.itineraryRepository = itineraryRepository;
+        this.contentRepository = contentRepository;
+        this.userService = userService;
     }
 
-    public Itinerary createItinerary(String name, String description, User creator, ContentCategory category, List<Content> pointsOfInterest) {
-        Itinerary itinerary = new Itinerary(name, description, creator,category, pointsOfInterest);
+    public Itinerary create(ItineraryDTO itDTO) {
+        //sets the creator as the authenticated user.
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = ((UserDetails) auth.getPrincipal()).getUsername();
+        User user = userService.getByUsername(username);
+
+        Itinerary itinerary = new Itinerary(itDTO.getName(), itDTO.getDescription(),user, itDTO.getCategory());
+        contentRepository.save(itinerary);
         return itineraryRepository.save(itinerary);
     }
 
-    public List<Itinerary> getALlItineraries() {
-        return (List<Itinerary>) itineraryRepository.findAll();
+    public Itinerary createVerified(ItineraryDTO itDTO) {
+        //sets the creator as the authenticated user.
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = ((UserDetails) auth.getPrincipal()).getUsername();
+        User user = userService.getByUsername(username);
+
+        Itinerary itinerary = new Itinerary(itDTO.getName(), itDTO.getDescription(),user, itDTO.getCategory());
+        itinerary.verify();
+        contentRepository.save(itinerary);
+        return itineraryRepository.save(itinerary);
     }
 
-    public Itinerary getItineraryById(Long id) {
-        return itineraryRepository.findById(id).orElse(null);
+    public List<Itinerary> getAll() {
+        List<Itinerary> itineraries = new ArrayList<>();
+        itineraryRepository.findAll().forEach(itineraries::add);
+        return itineraries;
     }
 
-    public void deleteItinerary(Long id) {
-        itineraryRepository.deleteById(id);
+    public Itinerary getItineraryById(Long id) throws ContentNotFoundException{
+        return itineraryRepository.findById(id).orElseThrow(ContentNotFoundException::new);
+    }
+
+    public void deleteItinerary(Long id) throws ContentNotFoundException {
+        if (!itineraryRepository.existsById(id)) {
+            throw new ContentNotFoundException();
+        } else {
+            contentRepository.deleteById(id);
+            itineraryRepository.deleteById(id);
+        }
+    }
+
+    public Itinerary addContent(Long idItinerary, Long idContent) throws ContentNotFoundException{
+        Itinerary itinerary = itineraryRepository.findById(idItinerary).orElse(null);
+        if (itinerary != null) {
+            itinerary.addContent(contentRepository.findByID(idContent).orElseThrow(ContentNotFoundException::new));
+            itineraryRepository.save(itinerary);
+            contentRepository.save(itinerary);
+            return itinerary;
+        } else {
+            throw new ContentNotFoundException();
+        }
     }
 }
